# bubuka_pro/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микросервиса",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/forgot/": {
            "post": {
                "tags": [
                    "Auth",
                    "Password"
                ],
                "summary": "Запрос на ссылку для сброса пароля",
                "operationId": "857804dcb05e54249b23121bc63a9daa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Адрес электронной почты",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Содержимое ответа",
                                            "type": "string",
                                            "example": "На почту отправлено письмо"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/registration/": {
            "post": {
                "tags": [
                    "Admin",
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "6de9501967d70761213492362bc621b2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Адрес электронной почты",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "is_admin": {
                                        "description": "Установка статуса. Админ или нет",
                                        "type": "boolean",
                                        "enum": [
                                            true,
                                            false
                                        ],
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус кода",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "description": "Сообщение ошибки",
                                            "properties": {
                                                "email": {
                                                    "description": "Возможные ошибки валидации поля email",
                                                    "type": "string",
                                                    "example": [
                                                        "The email field is required.",
                                                        "The email field must be a string",
                                                        "The email field must be a valid email address.",
                                                        "The email field must not be greater than 255 characters.",
                                                        "The email has already been taken."
                                                    ]
                                                },
                                                "is_admin": {
                                                    "description": "Возможная ошибка валидации поля is_admin",
                                                    "type": "string",
                                                    "example": "The is_admin field must be true or false"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "description": "Код ответа",
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/403"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "body": {
                        "properties": {
                            "users/user": {
                                "properties": {
                                    "id": {
                                        "description": "Id пользователя",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "description": "Почта пользователя",
                                        "type": "string",
                                        "example": "example@bubuka.info"
                                    },
                                    "is_admin": {
                                        "description": "Установка статуса. Админ или нет",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "deleted_at": {
                                        "description": "Дата удаления",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "created_at": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    },
                                    "updated_at": {
                                        "description": "Дата изменения",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "401": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": [
                            "Token not provided",
                            "The token has been blacklisted"
                        ]
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "401"
                    }
                },
                "type": "object"
            },
            "403": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Forbidden"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}